rework DataBlank to load ondemand further resource triples
in comparison to the "old" behavior, which sucks if you have
a complex and dense structure.

the new one lets you still use $a['bb:ccc'], but will load
further information if neccessary. that increases the
performance by some factors and avoids the problem to run
into cycles, as long as the user dont forces the class to.